Changing the value environments to partial functions.
This patch is the first in a series to change the type of value
environments from var -> val to var -> option val.
This has consequences on the development: All theorems become stronger
because they now state results about a model in which a variable can be
undefined. Some theorems (in particular correctness of delocation) must
be changed as they no longer hold in full generality.