Fix delocation by requiring definedness for additional parameters.
This is a fix for now, but it breaks the nice untyped setup we had
before the change of the environments from env to onv.
The problem is that additional paramters induce reads  function
applications -- which might get stuck now. In the old setup a variable
read could never get stuck.

There is an interesting relation between liveness and this problem. If
only live variables are added as additional parameters, it might look
like we can get around the definedness requirement, as any live variable
is guaranteed to be read later anyway. However, in the presence of
syscalls this is not true: the read may occur only after a syscall, i.e.
the translated program will crash before issuing the syscall -- and
hence is not equivalent.